bug 274394 - Exported product missing eclipse.ini

+++ org.eclipse.pde.build/src/org/eclipse/pde/internal/build/BrandingIron.java
@@ -123,9 +123,7 @@ public class BrandingIron implements IXMLConstants {
 		new File(target + "/MacOS").mkdirs(); //$NON-NLS-1$
 		new File(target + "/Resources").mkdirs(); //$NON-NLS-1$
 
-		String initialRoot = root + "/Launcher.app/Contents"; //$NON-NLS-1$
-		if (!new File(initialRoot).exists())
-			initialRoot = root + "/Eclipse.app/Contents"; //$NON-NLS-1$
+		String initialRoot = root + "/Eclipse.app/Contents"; //$NON-NLS-1$
 		copyMacLauncher(initialRoot, target);
 		String iconName = ""; //$NON-NLS-1$
 		File splashApp = new File(initialRoot, "Resources/Splash.app"); //$NON-NLS-1$
@@ -143,11 +141,13 @@ public class BrandingIron implements IXMLConstants {
 			brandMacSplash(initialRoot, target, iconName);
 		}
 
-		File rootFolder = new File(initialRoot);
-		rootFolder.delete();
-		if (rootFolder.exists()) {
-			//if the rootFolder still exists, its because there were other files that need to be moved over
-			moveContents(rootFolder, new File(target));
+		File rootFolder = getCanonicalFile(new File(initialRoot));
+		if (!rootFolder.equals(target)) {
+			rootFolder.delete();
+			if (rootFolder.exists()) {
+				//if the rootFolder still exists, its because there were other files that need to be moved over
+				moveContents(rootFolder, new File(target));
+			}
 		}
 		rootFolder.getParentFile().delete();
 	}
@@ -254,17 +254,17 @@ public class BrandingIron implements IXMLConstants {
 
 	private void copyMacLauncher(String initialRoot, String target) {
 		String targetLauncher = target + "/MacOS/"; //$NON-NLS-1$
-		File launcher = new File(initialRoot + "/MacOS/launcher"); //$NON-NLS-1$
-		File eclipseLauncher = new File(initialRoot + "/MacOS/eclipse"); //$NON-NLS-1$
+		File launcher = getCanonicalFile(new File(initialRoot + "/MacOS/launcher")); //$NON-NLS-1$
+		File eclipseLauncher = getCanonicalFile(new File(initialRoot + "/MacOS/eclipse")); //$NON-NLS-1$
+		File targetFile = getCanonicalFile(new File(targetLauncher, name));
 		if (!launcher.exists()) {
 			launcher = eclipseLauncher;
-		} else if (eclipseLauncher.exists()) {
+		} else if (eclipseLauncher.exists() && !targetFile.equals(eclipseLauncher)) {
 			//we may actually have both if exporting from the mac
 			eclipseLauncher.delete();
 		}
-		File targetFile = new File(targetLauncher, name);
 		try {
-			if (targetFile.getCanonicalFile().equals(launcher.getCanonicalFile())) {
+			if (targetFile.equals(launcher)) {
 				try {
 					//Force the executable bit on the exe because it has been lost when copying the file
 					Runtime.getRuntime().exec(new String[] {"chmod", "755", targetFile.getAbsolutePath()}); //$NON-NLS-1$ //$NON-NLS-2$
@@ -288,14 +288,22 @@ public class BrandingIron implements IXMLConstants {
 		launcher.getParentFile().delete();
 	}
 
+	private File getCanonicalFile(File file) {
+		try {
+			return file.getCanonicalFile();
+		} catch (IOException e) {
+			return file;
+		}
+	}
+
 	private void copyMacIni(String initialRoot, String target, String iconName) {
 		// 3 possibilities, in order of preference:
 		// rcp.app/Contents/MacOS/rcp.ini   		(targetFile)
 		// Eclipse.app/Contents/MacOS/rcp.ini		(brandedIni)
 		// Eclipse.app/Contents/MacOs/eclipse.ini	(ini)
-		File targetFile = new File(target, "/MacOS/" + name + ".ini"); //$NON-NLS-1$//$NON-NLS-2$
-		File brandedIni = new File(initialRoot, "/MacOS/" + name + ".ini"); //$NON-NLS-1$ //$NON-NLS-2$
-		File ini = new File(initialRoot, "/MacOS/eclipse.ini"); //$NON-NLS-1$
+		File targetFile = getCanonicalFile(new File(target, "/MacOS/" + name + ".ini")); //$NON-NLS-1$//$NON-NLS-2$
+		File brandedIni = getCanonicalFile(new File(initialRoot, "/MacOS/" + name + ".ini")); //$NON-NLS-1$ //$NON-NLS-2$
+		File ini = getCanonicalFile(new File(initialRoot, "/MacOS/eclipse.ini")); //$NON-NLS-1$
 
 		if (targetFile.exists()) {
 			//an ini already exists at the target, use that
